Increase log output readability
The current output of the Git step was using the log style of the controller,
which is configured to be machine consumable. However, the output should be
more aligned with the output that is created by other steps. Tools like Kaniko
or Buildpacks produce more human style output.

Replace `ctxlog` with `log` and simplify log output to be human friendly.

Add new end message that contains all relevant details, including HEAD info.